tiny

+++ hw5/2b/RNN_sentiment_analysis.py
@@ -100,7 +100,7 @@ for epoch in range(no_of_epochs):
 		x_input = glove_embeddings[x_input]
 		y_input = y_train[I_permutation[i:i+batch_size]]
 
-		data = Variable(torch.LongTensor(x_input)).cuda()
+		data = Variable(torch.FloatTensor(x_input)).cuda()
 		target = Variable(torch.FloatTensor(y_input)).cuda()
 
 		optimizer.zero_grad()
@@ -154,7 +154,7 @@ for epoch in range(no_of_epochs):
 		x_input = glove_embeddings[x_input]
 		y_input = y_test[I_permutation[i:i+batch_size]]
 
-		data = Variable(torch.LongTensor(x_input)).cuda()
+		data = Variable(torch.FloatTensor(x_input)).cuda()
 		target = Variable(torch.FloatTensor(y_input)).cuda()
 
 		with torch.no_grad():
